finance_app1/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-login==0.5.0
flask-migrate==3.1.0
flask-mail==0.9.1
python-dotenv==0.19.0
email-validator==1.1.3